break down trickle classes
