+++ server/trickle_integration.py
@@ -22,6 +22,11 @@ from pytrickle.frames import (
 from comfystream.pipeline import Pipeline
 from pytrickle.manager import BaseStreamManager
 
+# Import moved classes for backward compatibility
+from cleanup_manager import CleanupManager
+from trickle_stream_handler import TrickleStreamHandler
+from trickle_stream_manager import TrickleStreamManager
+
 logger = logging.getLogger(__name__)
 
 # Toggle to control whether audio frames are processed through the pipeline
@@ -40,60 +45,6 @@ PROCESS_AUDIO_THROUGH_PIPELINE = False
 # FrameProcessor has been moved to pytrickle.frame_processor
 # Import is now at the top of the file
 
-
-class CleanupManager:
-    """ComfyUI-specific cleanup management."""
-    
-    # Generic task cancellation moved to pytrickle.frames.StreamingUtils
-    cancel_task_with_timeout = StreamingUtils.cancel_task_with_timeout
-    
-    @staticmethod
-    async def cleanup_pipeline_resources(pipeline: Pipeline, request_id: str, timeout: float = 8.0) -> bool:
-        try:
-            async with asyncio.timeout(timeout):
-                try:
-                    await asyncio.wait_for(pipeline.client.cancel_running_prompts(), timeout=3.0)
-                except (asyncio.TimeoutError, Exception):
-                    pass
-                try:
-                    await asyncio.wait_for(pipeline.client.cleanup_queues(), timeout=2.0)
-                except (asyncio.TimeoutError, Exception):
-                    pass
-                return True
-        except (asyncio.TimeoutError, Exception):
-            return False
-    
-    @staticmethod
-    async def cleanup_memory(request_id: str, timeout: float = 10.0) -> bool:
-        try:
-            async with asyncio.timeout(timeout):
-                from comfystream import tensor_cache
-                def clear_caches():
-                    cleared = 0
-                    for cache in [tensor_cache.image_inputs, tensor_cache.audio_inputs]:
-                        while not cache.empty():
-                            try:
-                                cache.get_nowait()
-                                cleared += 1
-                            except:
-                                break
-                    # Clear text outputs cache as well
-                    async def clear_text_outputs():
-                        while not tensor_cache.text_outputs.empty():
-                            try:
-                                await tensor_cache.text_outputs.get()
-                            except:
-                                break
-                    try:
-                        import asyncio
-                        asyncio.create_task(clear_text_outputs())
-                    except:
-                        pass
-                    return cleared
-                await asyncio.to_thread(clear_caches)
-                return True
-        except (asyncio.TimeoutError, Exception):
-            return False
 
 class ComfyStreamTrickleProcessor(FrameConversionMixin):
     """Processes video frames through ComfyStream pipeline for trickle streaming."""
@@ -520,413 +471,3 @@ class ComfyStreamTrickleProcessor(FrameConversionMixin):
             fallback_frame = self.create_processed_frame(self.last_processed_frame.tensor, frame)
             return VideoOutput(fallback_frame, self.request_id)
         return VideoOutput(frame, self.request_id)
-
-
-class TrickleStreamHandler:
-    """Handles a complete trickle stream with ComfyStream integration."""
-    
-    def __init__(self, subscribe_url: str, publish_url: str, control_url: str, events_url: str,
-                 request_id: str, pipeline: Pipeline, width: int = 512, height: int = 512,
-                 data_url: Optional[str] = None, app_context: Optional[Dict] = None):
-        self.subscribe_url = subscribe_url
-        self.publish_url = publish_url
-        self.control_url = control_url
-        self.events_url = events_url
-        self.data_url = data_url
-        self.request_id = request_id
-        self.pipeline = pipeline
-        self.width = width
-        self.height = height
-        self.app_context = app_context or {}
-        
-        self.processor = ComfyStreamTrickleProcessor(pipeline, request_id)
-        
-        self.protocol = TrickleProtocol(
-            subscribe_url=subscribe_url, publish_url=publish_url, control_url=control_url,
-            events_url=events_url, data_url=data_url, width=width, height=height, error_callback=self._on_error
-        )
-        
-        self.client = TrickleClient(
-            protocol=self.protocol, frame_processor=self.processor.process_frame_sync,
-            control_handler=self._handle_control_message, error_callback=self._on_error
-        )
-        
-        self.control_subscriber = TrickleSubscriber(control_url, error_callback=self._on_error) if control_url and control_url.strip() else None
-        self.events_available = bool(events_url and events_url.strip())
-        
-        self.running_event = asyncio.Event()
-        self.shutdown_event = asyncio.Event()
-        self.error_event = asyncio.Event()
-        
-        self._task: Optional[asyncio.Task] = None
-        self._stats_task: Optional[asyncio.Task] = None
-        self._control_task: Optional[asyncio.Task] = None
-        self._critical_error_occurred = False
-        self._cleanup_lock = asyncio.Lock()
-
-    @property
-    def running(self) -> bool:
-        return self.running_event.is_set() and not self.shutdown_event.is_set() and not self.error_event.is_set()
-
-    async def _emit_monitoring_event(self, data: Dict[str, Any], event_type: str):
-        if not self.events_available:
-            return
-        try:
-            await self.client.protocol.emit_monitoring_event(data, event_type)
-        except Exception as e:
-            logger.warning(f"Failed to emit {event_type} event for {self.request_id}: {e}")
-    
-    async def _publish_data(self, text_data: str):
-        """Publish text data via the data channel."""
-        if not self.data_url:
-            return
-        try:
-            await self.client.publish_data(text_data)
-        except Exception as e:
-            logger.warning(f"Failed to publish data for {self.request_id}: {e}")
-    
-    async def _on_error(self, error_type: str, exception: Optional[Exception] = None):
-        if self.shutdown_event.is_set():
-            return
-        logger.error(f"Critical error for stream {self.request_id}: {error_type} - {exception}")
-        self._critical_error_occurred = True
-        if self.running:
-            self.error_event.set()
-            asyncio.create_task(self.stop())
-
-    async def _control_loop(self):
-        if not self.control_subscriber:
-            return
-        
-        keepalive_message = {"keep": "alive"}
-        try:
-            while not self.shutdown_event.is_set() and not self.error_event.is_set():
-                segment = await self.control_subscriber.next()
-                if not segment or segment.eos():
-                    break
-                
-                params_data = await segment.read()
-                if not params_data:
-                    continue
-                
-                try:
-                    params = json.loads(params_data.decode('utf-8'))
-                except (json.JSONDecodeError, UnicodeDecodeError) as e:
-                    logger.warning(f"Invalid control message: {e}")
-                    continue
-                if params == keepalive_message:
-                    continue
-                
-                await self._handle_control_message(params)
-        except asyncio.CancelledError:
-            raise
-        except Exception as e:
-            logger.error(f"Control loop error for stream {self.request_id}: {e}")
-            raise
-    
-    async def _handle_control_message(self, params: Dict[str, Any]):
-        try:
-            if "prompts" not in params:
-                return
-            
-            prompts = params["prompts"]
-            if isinstance(prompts, str):
-                try:
-                    prompts = json.loads(prompts)
-                except json.JSONDecodeError:
-                    return
-            
-            await self.pipeline.update_prompts(prompts)
-        except Exception as e:
-            logger.error(f"Error handling control message for {self.request_id}: {e}")
-            health_manager = self.app_context.get('health_manager')
-            if health_manager:
-                health_manager.set_error("Error handling control message for stream")
-    
-    async def _send_stats_periodically(self):
-        try:
-            while self.running:
-                try:
-                    stats = {
-                        "type": "stream_stats",
-                        "request_id": self.request_id,
-                        "timestamp": asyncio.get_event_loop().time(),
-                        "processor": {
-                            "frame_count": self.processor.frame_count,
-                            "pipeline_ready": self.processor.state.pipeline_ready,
-                            "buffer_stats": self.processor.frame_buffer.get_stats(),
-                            "last_processed_frame_available": self.processor.last_processed_frame is not None
-                        },
-                        "stream": {"running": self.running, "width": self.width, "height": self.height}
-                    }
-                    await self._emit_monitoring_event(stats, "stream_stats")
-                except Exception as e:
-                    logger.error(f"Error sending stats for {self.request_id}: {e}")
-                
-                # Wait for 20 seconds or until shutdown
-                try:
-                    await asyncio.wait_for(self.shutdown_event.wait(), timeout=20.0)
-                    break  # Shutdown requested
-                except asyncio.TimeoutError:
-                    pass  # Continue with next stats interval
-        except asyncio.CancelledError:
-            raise
-        except Exception as e:
-            logger.error(f"Stats monitoring error for {self.request_id}: {e}")
-    
-    async def start(self) -> bool:
-        if self.running:
-            return False
-        
-        try:
-            # Set up text output callback before starting processing
-            self.processor.set_text_output_callback(self._publish_data)
-            await self.processor.start_processing()
-            await self._emit_monitoring_event({"type": "stream_started", "request_id": self.request_id}, "stream_trace")
-            
-            pipeline_already_warmed = self.app_context.get("warm_pipeline", False)
-            
-            if pipeline_already_warmed:
-                await self.processor.set_pipeline_ready()
-            else:
-                try:
-                    await self.pipeline.warm_pipeline()
-                    if hasattr(self.pipeline, 'wait_for_first_processed_frame'):
-                        try:
-                            await self.pipeline.wait_for_first_processed_frame(timeout=30.0)
-                        except Exception:
-                            pass
-                except Exception as e:
-                    logger.error(f"Pipeline warmup failed: {e}")
-                await self.processor.set_pipeline_ready()
-            
-            self._task = asyncio.create_task(self.client.start(self.request_id))
-            self._task.add_done_callback(self._on_client_done)
-            self.running_event.set()
-            
-            try:
-                self._stats_task = asyncio.create_task(self._send_stats_periodically())
-            except Exception:
-                pass
-            
-            if self.control_url and self.control_url.strip():
-                try:
-                    self._control_task = asyncio.create_task(self._control_loop())
-                except Exception:
-                    pass
-            
-            return True
-        except Exception as e:
-            logger.error(f"Failed to start stream handler {self.request_id}: {e}")
-            self._set_final_state()
-            await self.processor.stop_processing()
-            return False
-    
-    def _on_client_done(self, task: asyncio.Task):
-        self.shutdown_event.set()
-        if task.exception():
-            logger.error(f"Client task for {self.request_id} finished with exception: {task.exception()}")
-            cleanup_task = asyncio.create_task(self.stop())
-            cleanup_task.add_done_callback(lambda t: None)
-        else:
-            stop_task = asyncio.create_task(self.stop())
-            stop_task.add_done_callback(lambda t: None)
-
-
-    async def stop(self, *, called_by_manager: bool = False) -> bool:
-        async with self._cleanup_lock:
-            self.shutdown_event.set()
-            
-            try:
-                try:
-                    await asyncio.wait_for(self.processor.stop_processing(), timeout=8.0)
-                except (asyncio.TimeoutError, Exception):
-                    pass
-                
-                try:
-                    await asyncio.wait_for(self.client.stop(), timeout=3.0)
-                except (asyncio.TimeoutError, Exception):
-                    pass
-                
-                await StreamingUtils.cancel_task_with_timeout(self._task, "Main task")
-                await StreamingUtils.cancel_task_with_timeout(self._stats_task, "Stats task")
-                await StreamingUtils.cancel_task_with_timeout(self._control_task, "Control loop")
-                
-                if self.control_subscriber:
-                    try:
-                        await self.control_subscriber.shutdown()
-                        await self.control_subscriber.close()
-                    except Exception:
-                        pass
-                
-                try:
-                    final_stats = {
-                        "type": "stream_stopped",
-                        "request_id": self.request_id,
-                        "timestamp": asyncio.get_event_loop().time(),
-                        "final_frame_count": self.processor.frame_count
-                    }
-                    await self._emit_monitoring_event(final_stats, "stream_trace")
-                except Exception:
-                    pass
-                
-                if not called_by_manager:
-                    await self._remove_from_manager_and_update_health()
-                
-                self._set_final_state()
-                return True
-                
-            except Exception as e:
-                logger.error(f"Error stopping stream handler {self.request_id}: {e}")
-                if not called_by_manager:
-                    await self._remove_from_manager_and_update_health(emergency=True)
-                self._set_final_state()
-                return False
-
-    def _set_final_state(self):
-        self.running_event.clear()
-        self.shutdown_event.set()
-        self.error_event.set()
-
-    async def _remove_from_manager_and_update_health(self, emergency: bool = False):
-        try:
-            stream_manager = self.app_context.get('stream_manager')
-            if stream_manager:
-                async with stream_manager.lock:
-                    if self.request_id in stream_manager.handlers:
-                        del stream_manager.handlers[self.request_id]
-                        health_manager = self.app_context.get('health_manager')
-                        if health_manager:
-                            stream_count = len(stream_manager.handlers)
-                            health_manager.update_trickle_streams(stream_count)
-                            if stream_count == 0:
-                                health_manager.clear_error()
-        except Exception as e:
-            logger.error(f"Error during cleanup for stream {self.request_id}: {e}")
-
-
-class TrickleStreamManager:
-    """ComfyUI-specific trickle stream manager that extends the base manager."""
-    
-    def __init__(self, app_context: Optional[Dict] = None):
-        # Extract health manager for parent class
-        health_manager = app_context.get('health_manager') if app_context else None
-        
-        self.handlers: Dict[str, TrickleStreamHandler] = {}
-        self.lock = asyncio.Lock()
-        self.app_context = app_context or {}
-        self.health_manager = health_manager
-    
-    async def create_stream(self, request_id: str, subscribe_url: str, publish_url: str,
-                          control_url: str, events_url: str, pipeline: Pipeline,
-                          width: int = 512, height: int = 512, data_url: Optional[str] = None) -> bool:
-        async with self.lock:
-            if request_id in self.handlers:
-                return False
-            
-            try:
-                # Clear error state if this is the first stream after error
-                if self.health_manager and self.health_manager.state == "ERROR":
-                    if len(self.handlers) == 0:
-                        self.health_manager.clear_error()
-                
-                handler = TrickleStreamHandler(
-                    subscribe_url=subscribe_url, publish_url=publish_url,
-                    control_url=control_url, events_url=events_url,
-                    request_id=request_id, pipeline=pipeline,
-                    width=width, height=height, data_url=data_url, app_context=self.app_context
-                )
-                
-                success = await handler.start()
-                if success:
-                    self.handlers[request_id] = handler
-                    self._update_health_manager()
-                    return True
-                else:
-                    return False
-            except Exception as e:
-                logger.error(f"Error creating stream {request_id}: {e}")
-                if self.health_manager:
-                    self.health_manager.set_error("Error creating trickle stream")
-                return False
-    
-    async def stop_stream(self, request_id: str) -> bool:
-        async with self.lock:
-            if request_id not in self.handlers:
-                return False
-            
-            handler = self.handlers[request_id]
-            success = await handler.stop(called_by_manager=True)
-            del self.handlers[request_id]
-            self._update_health_manager()
-            return success
-    
-    def _update_health_manager(self):
-        if self.health_manager:
-            stream_count = len(self.handlers)
-            # Use trickle-specific method if available (ComfyStreamHealthManager)
-            if hasattr(self.health_manager, 'update_trickle_streams'):
-                self.health_manager.update_trickle_streams(stream_count)
-            else:
-                self.health_manager.update_active_streams(stream_count)
-            if stream_count == 0 and self.health_manager.state == "ERROR":
-                self.health_manager.clear_error()
-    
-    def _get_stream_status_unlocked(self, request_id: str) -> Optional[Dict[str, Any]]:
-        """Get stream status without acquiring the lock (for internal use when lock is already held)."""
-        if request_id not in self.handlers:
-            return None
-        handler = self.handlers[request_id]
-        return {
-            'request_id': request_id, 'running': handler.running,
-            'subscribe_url': handler.subscribe_url, 'publish_url': handler.publish_url,
-            'control_url': handler.control_url, 'events_url': handler.events_url,
-            'events_available': handler.events_available, 'width': handler.width,
-            'height': handler.height, 'frame_count': handler.processor.frame_count,
-            'stats_monitoring_active': handler._stats_task is not None and not handler._stats_task.done(),
-            'pipeline_ready': handler.processor.state.pipeline_ready
-        }
-
-    async def get_stream_status(self, request_id: str) -> Optional[Dict[str, Any]]:
-        async with self.lock:
-            return self._get_stream_status_unlocked(request_id)
-    
-    async def list_streams(self) -> Dict[str, Dict[str, Any]]:
-        async with self.lock:
-            result = {}
-            for request_id in self.handlers:
-                status = self._get_stream_status_unlocked(request_id)
-                if status:
-                    result[request_id] = status
-            return result
-    
-    async def cleanup_all(self):
-        async with self.lock:
-            if not self.handlers:
-                return
-            
-            cleanup_tasks = []
-            for request_id in list(self.handlers.keys()):
-                cleanup_tasks.append(self._stop_stream_with_timeout(request_id))
-            
-            try:
-                await asyncio.wait_for(
-                    asyncio.gather(*cleanup_tasks, return_exceptions=True),
-                    timeout=15.0
-                )
-            except asyncio.TimeoutError:
-                logger.warning("Global cleanup timeout reached, forcing cleanup")
-            except Exception:
-                pass
-            
-            self.handlers.clear()
-            self._update_health_manager()
-    
-    async def _stop_stream_with_timeout(self, request_id: str) -> bool:
-        try:
-            if request_id in self.handlers:
-                handler = self.handlers[request_id]
-                return await asyncio.wait_for(handler.stop(called_by_manager=True), timeout=8.0)
-        except (asyncio.TimeoutError, Exception):
-            pass
-        return False
